Make additional map params remote

+++ app/src/main/java/pl/droidsonroids/droidsmap/model/Room.kt
@@ -11,17 +11,17 @@ data class Room(private val roomHeightPx: Float, private val roomWidthPx: Float,
 
     fun getRoomWidthMeters() = roomWidthPx * PX_TO_METERS_FACTOR
 
-    fun getRelativeCenterLatitude(translationDegrees: Float): Float {
-        val r = Math.sqrt(Math.pow(relativeCenterXPositionPx.toDouble(), 2.0) + Math.pow(relativeCenterYPositionPx.toDouble(), 2.0)).toFloat()
-        val phiRadians = Math.atan((relativeCenterYPositionPx / relativeCenterXPositionPx).toDouble()).toFloat()
-        val translatedPhiRadians = (phiRadians + Math.toRadians(translationDegrees.toDouble())).toFloat()
-        return (r.toDouble() * Math.cos(translatedPhiRadians.toDouble()) * PX_TO_METERS_FACTOR.toDouble() / LATITUDE_TO_METERS_FACTOR).toFloat()
+    fun getRelativeCenterLatitude(translationDegrees: Double): Double {
+        val r = Math.sqrt(Math.pow(relativeCenterXPositionPx.toDouble(), 2.0) + Math.pow(relativeCenterYPositionPx.toDouble(), 2.0))
+        val phiRadians = Math.atan((relativeCenterYPositionPx / relativeCenterXPositionPx).toDouble())
+        val translatedPhiRadians = (phiRadians + Math.toRadians(translationDegrees))
+        return (r * Math.cos(translatedPhiRadians) * PX_TO_METERS_FACTOR.toDouble() / LATITUDE_TO_METERS_FACTOR)
     }
 
-    fun getRelativeCenterLongitude(translationDegrees: Float): Float {
-        val r = Math.sqrt(Math.pow(relativeCenterXPositionPx.toDouble(), 2.0) + Math.pow(relativeCenterYPositionPx.toDouble(), 2.0)).toFloat()
-        val phiRadians = Math.atan((relativeCenterYPositionPx / relativeCenterXPositionPx).toDouble()).toFloat()
-        val translatedPhiRadians = (phiRadians + Math.toRadians(translationDegrees.toDouble())).toFloat()
-        return (r.toDouble() * Math.sin(translatedPhiRadians.toDouble()) * PX_TO_METERS_FACTOR.toDouble() / LONGITUDE_TO_METERS_FACTOR).toFloat()
+    fun getRelativeCenterLongitude(translationDegrees: Double): Double {
+        val r = Math.sqrt(Math.pow(relativeCenterXPositionPx.toDouble(), 2.0) + Math.pow(relativeCenterYPositionPx.toDouble(), 2.0))
+        val phiRadians = Math.atan((relativeCenterYPositionPx / relativeCenterXPositionPx).toDouble())
+        val translatedPhiRadians = (phiRadians + Math.toRadians(translationDegrees))
+        return (r * Math.sin(translatedPhiRadians) * PX_TO_METERS_FACTOR.toDouble() / LONGITUDE_TO_METERS_FACTOR)
     }
 }
